refactor debug scripts for improved readability and functionality

+++ oefeningen/debug4.py
@@ -22,12 +22,11 @@ def generate_sequence(current, previous, terms_left):
     return [current] + generate_sequence(next_number, current, terms_left - 1)
 
 def main():
-    start_value = np.uint8(1)
-    #start_value = 1  # The initial value
+    start_value = 1  # The initial value
     terms = 100  # The number of terms to generate
     
     # Generate the sequence starting with the given value
-    sequence = generate_sequence(start_value, np.uint8(0), terms)
+    sequence = generate_sequence(start_value, 0, terms)
     print(f"Generated sequence: {sequence}")
 
 if __name__ == "__main__":
